pyproject.toml
[project]
name = "run-agent"
version = "0.1.0"
description = "MiroFlow Project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "hydra-core>=1.3.2",
    "aiohttp>=3.12.15",
    "anthropic>=0.60.0",
    "e2b-code-interpreter>=1.2.1",
    "fastmcp>=2.10.6",
    "google-genai>=1.28.0",
    "markitdown-mcp>=0.0.1a3",
    "mcp>=1.12.2",
    "mutagen>=1.47.0",
    "openai==1.78.1",
    "rich>=14.1.0",
    "wikipedia>=1.4.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "ruff==0.8.0",
    "json5>=0.12.0",
    "pyyaml>=6.0.2",
    "tenacity>=8.2.3,<9.0.0",
    "tiktoken>=0.9.0",
    "datasets<4.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "filetype>=1.2.0",
    "Pillow",
    "pyzmq",
    "langgraph>=0.6.10",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "aiosqlite>=0.21.0",
    "yfinance>=0.2.66",
    "pandas-ta>=0.4.71b0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
